first

+++ modules/business/controllers/TripsController.php
@@ -5,9 +5,13 @@ namespace app\modules\business\controllers;
 use Yii;
 use app\modules\business\models\Trips;
 use app\modules\business\models\TripsSearch;
+use app\modules\business\models\MaterialTypes;
+use app\modules\business\models\BalanceSheet;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+
+
 
 /**
  * TripsController implements the CRUD actions for Trips model.
@@ -62,8 +66,21 @@ class TripsController extends Controller
     {
         $model = new Trips();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+        if (Yii::$app->request->post()) {
+                        $trip_count=$_POST['Trips']['trip_count'];
+                        for($i=1;$i<=$trip_count;$i++){
+                            $model = new Trips();
+                            $model->load(Yii::$app->request->post());
+                            $material_id=$model->material_id;
+                            $material_type=MaterialTypes::findOne($material_id);
+                            $model->measurement_type=$material_type->measurement_type;
+                            $model->date_of_travel = date("Y-m-d H:i:s",strtotime($model->date_of_travel));
+                            $model->buyer_amount=$model->buyer_amount_total-$model->vehicle_rent;
+                            $model->save();
+			}
+			
+            	return $this->redirect(['index']);
+			
         } else {
             return $this->render('create', [
                 'model' => $model,
@@ -79,10 +96,32 @@ class TripsController extends Controller
      */
     public function actionUpdate($id)
     {
+     $count = BalanceSheet::find()->where(['trip_id'=>$id,'status'=>'closed'])->count();
+     if($count > 0){
+            Yii::$app->session->setFlash('delete',"Cannot update, Bill generated already for this trip ");
+            return $this->redirect(['index']);
+        }   
+    
+    
         $model = $this->findModel($id);
+		if (Yii::$app->request->post()) {
+			$model->load(Yii::$app->request->post());
+			BalanceSheet::deleteAll('trip_id = :trip_id', [':trip_id' => $model->id]);
+			$model->trip_count = 1;
+			$material_id=$model->material_id;
+			$material_type=MaterialTypes::findOne($material_id);
+			$model->measurement_type=$material_type->measurement_type;
+			$model->date_of_travel = date("Y-m-d H:i:s",strtotime($model->date_of_travel));
+			if($model->save()){
+            	return $this->redirect(['view', 'id' => $model->id]);
+			}else{
+				return $this->render('view', [
+                'model' => $model,
+            ]);
+			}
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+        //if ($model->load(Yii::$app->request->post()) && $model->save()) {
+          //  return $this->redirect(['view', 'id' => $model->id]);
         } else {
             return $this->render('update', [
                 'model' => $model,
@@ -98,9 +137,19 @@ class TripsController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+        
+        $count = BalanceSheet::find()->where(['trip_id'=>$id,'status'=>'closed'])->count();
+        if($count==0){
+            
+            $this->findModel($id)->delete();
+            BalanceSheet::deleteAll('trip_id = :trip_id', [':trip_id' => $id]);
+            return $this->redirect(['index']);
+            }else{
+                Yii::$app->session->setFlash('delete',"Cannot delete, Bill generated already for this trip ");
+                return $this->redirect(['index']);
+            }
+            
+            
     }
 
     /**
